requirements.txt
certifi==2020.6.20
chardet==3.0.4
Django==2.2
django-haystack==3.0
elasticsearch==2.1.0
idna==2.10
install==1.3.4
mysqlclient==2.0.1
pkg-resources==0.0.0
pysolr==3.9.0
pytz==2020.1
requests==2.24.0
sqlparse==0.4.1
tika==1.24
urllib3==1.25.11
xapian-haystack==2.1.0
